requirements.txt
accelerate==1.6.0
aenum==3.1.16
aiohappyeyeballs==2.3.5
aiohttp==3.11.18
aiosignal==1.3.1
alembic==1.15.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
astor==0.8.1
async-timeout==4.0.3
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
backoff==2.2.1
banal==1.0.6
bcrypt==4.2.0
beautifulsoup4==4.12.3
bitsandbytes==0.45.5
blinker==1.9.0
blis==1.3.0
build==1.2.1
CacheControl==0.12.10
cachetools==5.4.0
cachy==0.3.0
catalogue==2.0.10
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.6.3
cleo==0.8.1
click==8.1.8
clikit==0.6.2
cloud-init==24.4
cloudpathlib==0.21.1
colorama==0.4.4
coloredlogs==15.0.1
command-not-found==0.3
confection==0.1.5
configobj==5.0.6
constantly==15.1.0
crashtest==0.3.1
cryptography==3.4.8
cymem==2.0.11
dataclasses-json==0.6.7
dataset==1.6.2
dbus-python==1.2.18
decorator==5.2.1
Deprecated==1.2.18
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
exceptiongroup==1.2.2
faiss-cpu==1.9.0
fastapi==0.112.0
filelock==3.15.4
Flask==3.1.0
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.33.0
google-search-results==2.4.2
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.65.4
h11==0.14.0
html2text==2024.2.26
html5lib==1.1
httpcore==1.0.5
httplib2==0.20.2
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.30.2
humanfriendly==10.0
hyperlink==21.0.0
idna==3.3
importlib_metadata==8.0.0
importlib_resources==6.4.0
incremental==21.3.0
itsdangerous==2.2.0
jeepney==0.7.1
jieba==0.42.1
Jinja2==3.1.6
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
kubernetes==30.1.0
langchain==0.3.24
langchain-chroma==0.2.3
langchain-community==0.2.12
langchain-core==0.3.56
langchain-huggingface==0.1.2
langchain-ollama==0.1.1
langchain-openai==0.3.0
langchain-text-splitters==0.3.8
langchainhub==0.1.21
langcodes==3.5.0
langgraph==0.2.62
langgraph-checkpoint==2.0.9
langgraph-sdk==0.1.51
langsmith==0.1.147
language_data==1.3.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
line-bot-sdk==3.17.1
lockfile==0.12.2
Mako==1.3.10
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.21.3
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.0.5
murmurhash==1.0.12
mypy-extensions==1.0.0
narwhals==1.28.0
netifaces==0.11.0
networkx==3.3
numexpr==2.10.2
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
ollama==0.3.1
onnxruntime==1.18.1
openai==1.59.6
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
opt-einsum==3.3.0
orjson==3.10.7
overrides==7.7.0
packaging==24.1
paddlepaddle==3.0.0
pandas==2.2.3
pastel==0.2.1
pexpect==4.8.0
pillow==10.4.0
pkginfo==1.8.2
platformdirs==2.5.1
playwright==1.46.0
poetry==1.1.12
poetry-core==1.0.7
posthog==3.5.0
preshed==3.0.9
propcache==0.3.1
protobuf==4.25.4
psutil==7.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pyarrow==19.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.1
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pyee==11.1.0
Pygments==2.18.0
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pylev==1.2.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyPika==0.48.9
pyproject_hooks==1.1.0
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-magic==0.4.24
pytz==2022.1
PyYAML==6.0.2
rank-bm25==0.2.2
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rsa==4.9
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.0
SecretStorage==3.3.1
sentence-transformers==3.0.1
sentencepiece==0.2.0
service-identity==18.1.0
shellingham==1.5.4
six==1.16.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sos==4.7.2
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==1.4.54
srsly==2.5.1
ssh-import-id==5.11
starlette==0.37.2
streamlit==1.42.2
sympy==1.13.2
systemd-python==234
tenacity==8.5.0
text-generation==0.7.0
texttable==1.6.4
thinc==8.3.6
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.9.2
torch==2.4.0
tornado==6.4.2
tqdm==4.66.5
transformers==4.51.1
triton==3.0.0
Twisted==22.1.0
typer==0.12.3
types-requests==2.32.0.20250328
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.3.0
uvicorn==0.30.6
uvloop==0.19.0
virtualenv==20.13.0+ds
wadllib==1.3.6
wasabi==1.1.3
watchdog==6.0.0
watchfiles==0.23.0
weasel==0.4.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.1.3
wrapt==1.16.0
xkit==0.0.0
yarl==1.20.0
zipp==1.0.0
zope.interface==5.4.0
zstandard==0.23.0
